requirements.txt
asgiref==3.8.0
Django==5.0.3
django-cors-headers==4.3.1
django-filter==24.1
django-rest-framework==0.1.0
djangorestframework==3.15.0
djangorestframework-simplejwt==5.3.1
mysqlclient==2.2.4
pillow==10.2.0
PyJWT==2.8.0
sqlparse==0.4.4
tzdata==2024.1
